float again?

+++ models/modelUtils.py
@@ -33,7 +33,7 @@ def regression_loss(output, target, num_classes, alpha=0.9):
     pose_loss = F.mse_loss(output_pose, target_pose).float()
 
     # Combine them with the alpha values
-    total_loss = alpha * class_loss + (1 - alpha) * pose_loss
+    total_loss = float(alpha * class_loss + (1 - alpha) * pose_loss)
 
     return total_loss
